PD#152914: sensor: fixed sensors defects
driver defect clean up:

 #380~381 390~392 417 ltr558_pls.c
 #388~389 397 401 mm3a310.c
 #402 mir3da_cust.c
 #404 stk831x.c
 #405 stk8313.c
 #426 lsm303d.c

+++ drivers/amlogic/input/sensor/lsm303d.c
@@ -727,7 +727,7 @@ static irqreturn_t lsm303d_isr2(int irq, void *dev)
 static void lsm303d_interrupt_catch(struct lsm303d_status *stat, int pin)
 {
 	u8 buf[2];
-	u8 val;
+	int val;
 
 	if (atomic_read(&stat->interrupt->gen1_pin1.enable) == 1) {
 		buf[0] = status_registers.int_gen1_src.address;
